Refactor: Extract wargame engine logic to hook
This commit extracts the complex logic from NeuralNetworkCanvas into a new hook, useWargameEngine. This improves maintainability, testability, and performance. It also includes updates to other components for better type safety and cleanup.
