+++ src/app/_components/terminal/data-stream-display.tsx
@@ -34,7 +34,7 @@ export function DataStreamDisplay({
   const [entries, setEntries] = useState<DataEntry[]>([]);
   const [isPaused, setIsPaused] = useState(false);
   const streamRef = useRef<HTMLDivElement>(null);
-  const intervalRef = useRef<NodeJS.Timeout>();
+  const intervalRef = useRef<NodeJS.Timeout | undefined>(undefined);
 
   // Simulate data stream
   useEffect(() => {
